+++ src/AzureAuth/Commands/CommandAad.cs
@@ -7,7 +7,6 @@ namespace Microsoft.Authentication.AzureAuth.Commands
     using System.Collections.Generic;
     using System.IO.Abstractions;
     using System.Linq;
-    using System.Threading.Tasks;
 
     using McMaster.Extensions.CommandLineUtils;
 
@@ -297,7 +296,7 @@ Allowed values: [all, web, devicecode]";
         /// <returns>
         /// The error code: 0 is normal execution, and the rest means errors during execution.
         /// </returns>
-        public int OnExecute()
+        public int OnExecute(IPublicClientAuth publicClientAuth)
         {
             if (!this.EvaluateOptions())
             {
@@ -314,17 +313,7 @@ Allowed values: [all, web, devicecode]";
             // Small bug in Lasso - Add does not accept a null IEnumerable here.
             this.eventData.Add("settings_scopes", this.authSettings.Scopes ?? new List<string>());
 
-            if (this.env.InteractiveAuthDisabled())
-            {
-                this.eventData.Add(EnvVars.CorextNonInteractive, this.env.Get(EnvVars.CorextNonInteractive));
-                this.eventData.Add(EnvVars.NoUser, this.env.Get(EnvVars.NoUser));
-                this.logger.LogWarning($"Interactive authentication is disabled.");
-#if PlatformWindows
-                this.logger.LogWarning($"Only Integrated Windows Authentication will be attempted.");
-#endif
-            }
-
-            return this.ClearCache ? this.ClearLocalCache() : this.GetToken();
+            return this.ClearCache ? this.ClearLocalCache() : this.GetToken(publicClientAuth);
         }
 
         private bool ValidateOptions()
@@ -377,38 +366,23 @@ Allowed values: [all, web, devicecode]";
             return 0;
         }
 
-        private int GetToken()
+        private int GetToken(IPublicClientAuth publicClientAuth)
         {
             try
             {
-                var results = TokenFetcher.AccessToken(
-                    logger: this.logger,
-                    client: new Guid(this.authSettings.Client),
-                    tenant: new Guid(this.authSettings.Tenant),
-                    scopes: this.authSettings.Scopes,
-                    mode: this.AuthModes.Combine().PreventInteractionIfNeeded(this.env),
+                TokenResult tokenResult = publicClientAuth.Token(
+                    authParams: new AuthParameters(this.authSettings.Client, this.authSettings.Tenant, this.authSettings.Scopes),
+                    authModes: this.AuthModes,
                     domain: this.authSettings.Domain,
-                    prompt: AzureAuth.PromptHint.Prefixed(this.authSettings.PromptHint),
-                    timeout: TimeSpan.FromMinutes(this.Timeout));
+                    prompt: this.authSettings.PromptHint,
+                    timeout: TimeSpan.FromMinutes(this.Timeout),
+                    this.eventData);
 
-                var errors = results.Attempts.SelectMany(attempt => attempt.Errors).ToArray();
-                this.eventData.Add("error_count", errors.Length);
-                this.eventData.Add("authflow_count", results.Attempts.Count);
-
-                // Send custom telemetry events for each authflow result.
-                results.Attempts.SendTelemetry(this.telemetryService);
-
-                var success = results.Success;
-                if (success == null)
+                if (tokenResult == null)
                 {
                     this.logger.LogError("Authentication failed. Re-run with '--verbosity debug' to get see more info.");
                     return 1;
                 }
-
-                var tokenResult = success.TokenResult;
-                this.eventData.Add("silent", tokenResult.IsSilent);
-                this.eventData.Add("sid", tokenResult.SID);
-                this.eventData.Add("succeeded_mode", success.AuthFlowName);
 
                 switch (this.Output)
                 {