+++ src/AzureAuth/Program.cs
@@ -4,10 +4,15 @@
 namespace Microsoft.Authentication.AzureAuth
 {
     using System;
+    using System.Runtime.InteropServices;
     using System.Text;
 
     using McMaster.Extensions.CommandLineUtils;
+
     using Microsoft.Authentication.AzureAuth.Commands;
+    using Microsoft.Authentication.MSALWrapper;
+    using Microsoft.Extensions.DependencyInjection;
+    using Microsoft.Extensions.Logging;
     using Microsoft.Office.Lasso;
     using Microsoft.Office.Lasso.Telemetry;
 
@@ -31,7 +36,7 @@ namespace Microsoft.Authentication.AzureAuth
             TelemetryOutput backend = TelemetryOutput.StandardOut;
 
             // We will only send telemetry if we are given an ingestion token via
-            // environment variable. We *have* to do this here, rather than in
+            // environment variable or registry. We *have* to do this here, rather than in
             // `CommandMain.OnExecute`, because Lasso doesn't have a way of allowing a
             // `CommandLineApplication` to dynamically set telemetry configuration. Even if
             // it did, we can't guarantee that a failure in command line parsing wouldn't
@@ -39,18 +44,38 @@ namespace Microsoft.Authentication.AzureAuth
             //
             // To disable telemetry a user need only leave this environment variable unset. It's off by default.
             string applicationInsightsIngestionToken = Environment.GetEnvironmentVariable(EnvVars.ApplicationInsightsIngestionTokenEnvVar);
+
+            if (string.IsNullOrEmpty(applicationInsightsIngestionToken))
+            {
+                if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
+                {
+                    applicationInsightsIngestionToken = Win32.Registry.GetValue(
+                        EnvVars.ApplicationInsightsIngestionTokenRegKeyPath,
+                        EnvVars.ApplicationInsightsIngestionTokenRegKeyName,
+                        null) as string;
+                }
+            }
+
             if (!string.IsNullOrEmpty(applicationInsightsIngestionToken))
             {
                 ingestionToken = applicationInsightsIngestionToken;
                 backend = TelemetryOutput.ApplicationInsights;
             }
 
+            var envVarsToCollect = new[]
+            {
+                Ado.Constants.SystemDefinitionId,
+                EnvVars.CloudBuild,
+                EnvVars.NoUser,
+                EnvVars.CorextNonInteractive,
+            };
+
             TelemetryConfig telemetryConfig = new TelemetryConfig(
                 eventNamespace: "azureauth",
                 backend: backend,
                 ingestionToken: ingestionToken,
                 useAsync: true,
-                envVarsToCollect: new[] { "SYSTEM_DEFINITIONID", "QBUILD_DISTRIBUTED" },
+                envVarsToCollect: envVarsToCollect,
                 hideAlias: true,
                 hideMachineName: true);
 
@@ -64,7 +89,15 @@ namespace Microsoft.Authentication.AzureAuth
                 sendCommandEvents: true,
                 minStderrLoglevel: stdErrLogLevel);
 
-            new Lasso(app, options).Execute(args);
+            var loggerFactory = new NLog.Extensions.Logging.NLogLoggerFactory();
+            var logger = loggerFactory.CreateLogger("AzureAuth");
+
+            IServiceCollection services = new ServiceCollection();
+            services.AddSingleton<IMsalWrapper, MsalWrapper>();
+            services.AddSingleton<IPublicClientAuth, PublicClientAuth>();
+            services.AddSingleton<ILogger>(logger);
+
+            new Lasso(app, options, services).Execute(args);
         }
     }
 }